// lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      analysis_history: {
        Row: {
          analysis_result: Json | null
          created_at: string
          id: string
          user_id: string
          video_storage_path: string | null
          video_url: string | null
        }
        Insert: {
          analysis_result?: Json | null
          created_at?: string
          id?: string
          user_id: string
          video_storage_path?: string | null
          video_url?: string | null
        }
        Update: {
          analysis_result?: Json | null
          created_at?: string
          id?: string
          user_id?: string
          video_storage_path?: string | null
          video_url?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "analysis_history_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      users: {
        Row: {
          created_at: string
          email: string
          id: string
          period_reset_at: string
          plan: string
          stripe_customer_id: string | null
          stripe_subscription_id: string | null
          updated_at: string
          usage_count: number
        }
        Insert: {
          created_at?: string
          email: string
          id: string
          period_reset_at?: string
          plan?: string
          stripe_customer_id?: string | null
          stripe_subscription_id?: string | null
          updated_at?: string
          usage_count?: number
        }
        Update: {
          created_at?: string
          email?: string
          id?: string
          period_reset_at?: string
          plan?: string
          stripe_customer_id?: string | null
          stripe_subscription_id?: string | null
          updated_at?: string
          usage_count?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}